refactor(actions/tasks): co-locate project actions in subfolders
